+++ simple_deploy/management/commands/platform_sh/tests/e2e_tests/__init__.py
@@ -1,1 +1,0 @@
-# This file is needed to support importing helper functions from utils/.